Improve efficiency
By using a while loop over the list of arguments and appending to that
list, there is no need to run a subshell each time the user navigates
to another directory.

One side effect of this patch is that pathmenu no longer accepts multiple
relative arguments for directories to begin in.  This is because such
relative references would become irrelevant once the first cd takes place
(remember it's not in a subshell any more!).

A solution to the above might be to identify which directory arguments
are relative and prefix those with $PWD before entering the initial
while loop where the cd commands takes place.  Something like this:

    for arg do
        shift
        case $arg in
            /*) set -- "$@" "$arg"      ;;  # absolute
            *)  set -- "$@" "$PWD/$arg" ;;  # relative
        esac
    done