Globs are now specified by repeated -p options
This is slightly less convenient in some cases but overall
much more powerful.  You can still achieve the effect of the
options that have now been removed:

    pathmenu -a   => pathmenu -p'.*' -p'*'
    pathmenu -A   => pathmenu -p'.[^.]' -p'..?*' -p'*'
    pathmenu -Aa  => pathmenu -p.. -p. -p'*'
    pathmenu -F   => pathmenu -p'*/' -p'*'
    pathmenu -aF  => pathmenu -p'.*/' -p'.*' -p'*/' -p'*'
    pathmenu -AF  => pathmenu -p'.[^.]*/' -p'..?*' -p'*/' -p'*'
    pathmenu -AaF => pathmenu -p../ -p./ -p'*/' -p'*'

What is lost in convenience is gained in flexibility, because now it is
possible to specify arbitrary globs!  For example to suggest only PNG
and JPG file names:

    pathmenu -p \*.png -p \*.jpg

I may restore the -AadF options in a later patch for convenience and
backwards compatibility—though their effects are in conflict so I’d
need to find a way to resolve these (probably just ignore -AadF if any
-p given).