Accept options -dn
With -d suggestions are limited to directories only.

With -n suggestions are not limited in any way except that they must not
be the empty string.  This means you can suggest non-existent files by
providing -p PATTERN like so:

    pathmenu -n -p foo -p bar

Even if foo and bar do not exist, they will still be listed as suggestions
for completion in dmenu (or fzf or whatever).  But

    pathmenu -n -p ''

will list nothing whatsoever.