Stop shellcheck from whining again
THIS IS SAFE

Curiously it complains about

    [ -"$TEST" "$match" ]

but not

    test -"$TEST" "$match"

which is equivalent to my understanding.  Maybe it’s more lenient
with test(1) than with [(1) because test could be the name of a
function?  At any rate, adding the directive

    # shellcheck disable=SC1072
    [ -"$TEST" "$match" ] || continue

did not work, strangely.  Maybe I’m doing it wrong.  And no, it does
not like "-$TEST" any better than -"$TEST".